Sensor should run once per project.
So it runs only once. In sonar we have project, and project can have modules. If `Sensor` is not global it will run on every module. Important for `Sonar Scanner for Net` where every `csproj` is a module. Because we are applying metrics from one file, taken for one repository, it make sense to mark this plugin global.